servers/comfyui/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
python-dotenv
forex-python
requests
pytz
python-dateutil
pillow